start of http refactor
